requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
Pillow==10.4.0
boto3==1.35.0
django-storages==1.14.4
sentry-sdk==2.12.0
structlog==24.4.0
django-structlog==8.1.0
python-decouple==3.8
django-filter==24.3
drf-spectacular==0.27.2
openai==1.40.0
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
Faker==26.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
whitenoise==6.7.0
gunicorn==22.0.0
uvicorn==0.30.6
asgiref==3.8.1
